Restructure texture invalidation chaining
The original texture cache will invalidate any section that overlaps the
faulting page, as well as any section that overlaps an invalidated
section.

This means that it is possible to have a situation where a single fault
triggers many invalidations, killing performance. As a result,
conservative protection mode was set as the default, which reduces the
number of protected pages per section but introduces possible
texture corruption or even crashes in very rare cases.

This commit re-enables full range protection (instead of conservative
protection) and re-implements the way texture invalidation chaining is
done, to avoid the corresponding performance loss.