Refactor section refcounting/memory management
Instead of requiring manual management, sections now get automatically
handled and destroyed when they are no longer in use. This should remove
potential memory leaks and similar issues.

Note that framebuffer sections are handled differently, and thus still
require a manual destroy() call. This is because these are expensive
sections to allocate, and as such will be kept around for a while in the
hopes that they will be reused soon.